+++ src/Service/PaymentMethodService.php
@@ -23,6 +23,7 @@ use Kiener\MolliePayments\Handler\Method\KlarnaPayLaterPayment;
 use Kiener\MolliePayments\Handler\Method\KlarnaPayNowPayment;
 use Kiener\MolliePayments\Handler\Method\KlarnaSliceItPayment;
 use Kiener\MolliePayments\Handler\Method\MbWayPayment;
+use Kiener\MolliePayments\Handler\Method\MultibancoPayment;
 use Kiener\MolliePayments\Handler\Method\MyBankPayment;
 use Kiener\MolliePayments\Handler\Method\PayByBankPayment;
 use Kiener\MolliePayments\Handler\Method\PayconiqPayment;
@@ -462,7 +463,8 @@ class PaymentMethodService
             RivertyPayment::class,
             SatispayPayment::class,
             PayByBankPayment::class,
-            MbWayPayment::class
+            MbWayPayment::class,
+            MultibancoPayment::class,
             // IngHomePayPayment::class, // not allowed anymore
             // DirectDebitPayment::class, // only allowed when updating subsriptions, aka => not allowed anymore
         ];