+++ tests/Cypress/cypress/e2e/storefront/checkout/checkout-success.cy.js
@@ -61,6 +61,8 @@ const payments = [
     {caseId: 'C3713512', key: 'satispay', name: 'Satispay', sanity: false},
     {caseId: 'C4212005', key: 'paybybank', name: 'Pay by Bank', sanity: false},
     {caseId: '', key: 'mbway', name: 'MB Way', sanity: false},
+    {caseId: '', key: 'multibanco', name: 'Multibanco', sanity: false},
+
     // unfortunately address and product prices need to match, so we cannot do in3 automatically for now
     // {caseId: '', key: 'in3', name: 'in3'},
 ];